pyproject.toml
[tool.poetry]
name = "ai-course-2024-1"
version = "0.1.0"
description = ""
authors = ["yenslife <77geo5rge6@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.57.4"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
numpy = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
